Create base class for SheetCommand
